+++ src/app/admin/admin-routing.module.ts
@@ -5,7 +5,6 @@ import { InventoryComponent } from './inventory/inventory.component';
 import { EventComponent } from './event/event.component';
 import { FaqComponent } from './faq/faq.component';
 import { SupplierComponent } from './supplier/supplier.component';
-import { ReportComponent } from './report/report.component';
 import { BlacklistComponent } from './blacklist/blacklist.component';
 import { VatComponent } from './vat/vat.component';
 import { EmployeeComponent } from './employee/employee.component';
@@ -38,7 +37,6 @@ const routes: Routes = [
   { path: 'event', component: EventComponent },
   { path: 'faq', component: FaqComponent },
   { path: 'supplier', component: SupplierComponent },
-  { path: 'report', component: ReportComponent },
   { path: 'writeoff', component: WriteoffComponent},
   { path: 'blacklist', component: BlacklistComponent },
   { path: 'vat', component: VatComponent},
