--- src/app/admin/report/report.component.html
@@ -1,86 +1,0 @@
-<br>
-<br>
-<h2>Generate reports</h2>
-<div>
-  <!-- <button type="button" class="btn btn-primary" (click)="generateReport()">Generate Report</button> -->
-  <button type="button" class="btn btn-primary" style="margin: 2px;" (click)="showModal('WINES')">Wines</button>
-  <!-- <button type="button" class="btn btn-primary" style="margin: 2px;" (click)="showDateModal('EVENTS')">Events</button> -->
-  <!-- <button type="button" class="btn btn-primary" style="margin: 2px;" (click)="showToastr('Sales')">Sales</button> -->
-  <!-- <button type="button" class="btn btn-primary" style="margin: 2px;" (click)="showDateModal('REFUNDS')">Refunds</button> -->
-  <!-- <button type="button" class="btn btn-primary" style="margin: 2px;" (click)="showModal('SUPPLIER ORDER')">Supplier Order</button> -->
-  <button type="button" class="btn btn-primary" style="margin: 2px;" (click)="showModal('BLACKLIST')">Blacklist</button>
-  <button type="button" class="btn btn-primary" style="margin: 2px;" (click)="showModal('INVENTORY')">Inventory</button>
-</div>
-
-<!-- Events Report Modal -->
-<div class="modal" tabindex="-1" [ngClass]="{'show': showRefundsModal}"
-  [ngStyle]="{'display': showRefundsModal ? 'block' : 'none'}">
-  <div class="modal-dialog">
-    <div class="modal-content">
-      <div class="modal-header">
-        <h1>{{ currentReportType }}</h1>        
-        <button type="button" class="btn-close" aria-label="Close" (click)="closeRefundsModal()"></button>
-      </div>
-      <div class="modal-body">
-        <h5 class="modal-title">Please select the dates you wish to create a {{ currentReportType?.toLowerCase() }} report for</h5>
-        <div class="form-group">
-          <label for="beginDateInput">Begin Date:</label>
-          <input type="date" id="beginDateInput" [(ngModel)]="beginDate">
-        </div>
-        <div class="form-group">
-          <label for="endDateInput">End Date:</label>
-          <input type="date" id="endDateInput" [(ngModel)]="endDate">
-        </div>
-      </div>
-      <div class="modal-footer">
-        <button type="button" class="btn btn-primary" (click)="OpenDateReports(); AddAuditLog('Report: ' + currentReportType + ' Opened')">Open Report</button>
-        <button type="button" class="btn btn-primary" (click)="DownloadDateReports(); AddAuditLog('Report: ' + currentReportType + ' Downloaded')">Download Report</button>
-      </div>
-    </div>
-  </div>
-</div>
-
-  <!-- Sales Report Modal -->
-<div class="modal" tabindex="-1" [ngClass]="{'show': showSalesModal}" [ngStyle]="{'display': showSalesModal ? 'block' : 'none'}">
-  <div class="modal-dialog">
-    <div class="modal-content">
-      <div class="modal-header">
-        <h5 class="modal-title">Please select the dates for the Sales Report</h5>
-        <button type="button" class="btn-close" aria-label="Close" (click)="closeSalesModal()"></button>
-      </div>
-      <div class="modal-body">
-        <div class="form-group">
-          <label for="beginDateInput">Begin Date:</label>
-          <input type="date" id="beginDateInput" [(ngModel)]="beginDate">
-        </div>
-        <div class="form-group">
-          <label for="endDateInput">End Date:</label>
-          <input type="date" id="endDateInput" [(ngModel)]="endDate">
-        </div>
-      </div>
-      <div class="modal-footer">
-        <button type="button" class="btn btn-secondary" (click)="closeSalesModal()">Cancel</button>
-        <button type="button" class="btn btn-primary" (click)="generateSalesChartAndPDF()">Generate Report</button>
-      </div>
-    </div>
-  </div>
-</div>
-
-<!-- Blacklist Report Modal -->
-<div class="modal" tabindex="-1" [ngClass]="{'show': showBlacklistModal}" [ngStyle]="{'display': showBlacklistModal ? 'block' : 'none'}">
-  <div class="modal-dialog">
-    <div class="modal-content">
-      <div class="modal-header">
-        <h1>{{ currentReportType }}</h1>        
-        <button type="button" class="btn-close" aria-label="Close" (click)="closeBlacklistModal()"></button>
-      </div>
-      <div class="modal-body">
-        <h5 class="modal-title">Please select how you want to view the {{ currentReportType?.toLowerCase() }} report</h5>
-      </div>      
-      <div class="modal-footer">
-        <button type="button" class="btn btn-primary" (click)="OpenReports(); AddAuditLog('Report: ' + currentReportType + ' Opened')">Open Report</button>
-        <button type="button" class="btn btn-primary" (click)="DownloadReports(); AddAuditLog('Report: ' + currentReportType + ' Downloaded')">Download Report</button>
-      </div>
-    </div>
-  </div>
-</div>
