+++ src/app/admin/wine/wine.component.ts
@@ -16,6 +16,9 @@ import { DataServiceService } from 'src/app/customer/services/data-service.servi
 import { AuditlogService } from '../services/auditlog.service';
 import { CustomersService } from '../services/customers.service';
 
+import { PdfService } from '../services/pdf.service';
+import { ReportService } from '../services/report.service';
+
 
 @Component({
   selector: 'app-wine',
@@ -37,7 +40,8 @@ export class WineComponent implements OnInit {
 
   constructor(private toastr: ToastrService, private discountService: DiscountService, private router: Router, private wineService: WineService
     , private winetypeService: WinetypeService, private varietalService: VarietalService, private changeDetector: ChangeDetectorRef
-    , private customerService: CustomersService, private auditLogService: AuditlogService, private dataService: DataServiceService) { }
+    , private customerService: CustomersService, private auditLogService: AuditlogService, private dataService: DataServiceService
+    ,private pdfService: PdfService,) { }
 
   ngOnInit(): void {
     this.loadVarietals();
@@ -448,6 +452,86 @@ filterWines(): void {
     this.pageSize = newSize;
     this.filterWines(); // Re-filter wines without resetting allWines
   }
+
+  currentReportType: 'REFUNDS' | 'EVENTS' | 'BLACKLIST' | 'INVENTORY' | 'SUPPLIER ORDER' | 'WINES' | null = null;
+  showBlacklistModal: boolean = false;
+
+  showModal(reportType: 'BLACKLIST' | 'INVENTORY' | 'SUPPLIER ORDER' | 'WINES'): void {
+    this.currentReportType = reportType;
+    this.showBlacklistModal = true;
+  }
+  closeBlacklistModal() {
+    this.showBlacklistModal = false;
+    this.currentReportType = null;
+  }
+
+  OpenReports(): void {
+   
+       if (this.currentReportType === 'WINES') {
+      this.ViewWines();
+    }
+  }
+
+  DownloadReports(): void {
+    if (this.currentReportType === 'WINES') {
+      this.generateWineReportpdf();
+    }
+  }
+
+
+  async ViewWines(): Promise<void> {
+    try {
+      const currentDate = this.getCurrentDateFormatted(); // Get the current date
+      console.log(this.wines)
+      const pdfBlob = await this.pdfService.generateWinesReport(this.wines, currentDate);
+
+      const resolvedPdfBlob = await pdfBlob;
+
+      // Create a Blob URL and open it in a new tab
+      const blobUrl = URL.createObjectURL(resolvedPdfBlob);
+      const newTab = window.open(blobUrl, '_blank');
+      if (!newTab) {
+        console.error('Failed to open new tab for PDF');
+      } else {
+        console.error('Received undefined or invalid Wine data');
+      }
+    } catch (error) {
+      console.error('Error generating Wine report:', error);
+    }
+  }
+
+
+  async generateWineReportpdf() {
+    try {
+      let result: Wine[] | undefined = await this.wineService.getWines();
+      console.log('Result:', result); // Add this line
+      if (result !== undefined) {
+        this.wines = result;
+        let currentDate = this.getCurrentDateFormatted();
+        this.pdfService.generateWinesReportpdf(this.wines, currentDate);
+        // Do any additional processing or actions here if needed
+      } else {
+        console.error('Received undefined or invalid wine data');
+        this.toastr.error('Error, failed to download Wines Report', 'Wines Report');
+        // Handle the case where the returned data is undefined or invalid
+      }
+    } catch (error) {
+      console.error('Error fetching wine data:', error);
+      this.toastr.error('Error, failed to retrieve Wine Data', 'Wines Report');
+      // Handle error if needed
+    }
+  }
+
+
+  getCurrentDateFormatted(): string {
+    const today = new Date();
+    const day = today.getDate().toString().padStart(2, '0');
+    const month = (today.getMonth() + 1).toString().padStart(2, '0');
+    const year = today.getFullYear();
+    return `${day}/${month}/${year}`;
+  }
+
+
